requirements.txt
streamlit
google-generativeai
python-dotenv
chromadb
faiss-cpu
